rename device fingerprint component and move it to the webflow package
