+++ support/cas-server-support-trusted-mfa/src/main/java/org/apereo/cas/trusted/web/flow/DeviceFingerprintStrategy.java
@@ -1,23 +1,23 @@
-package org.apereo.cas.trusted.fingerprint;
+package org.apereo.cas.trusted.web.flow;
 
 import org.springframework.webflow.execution.RequestContext;
 
 import javax.annotation.Nonnull;
 
 /**
- * Interface for generating a device fingerprint for usage within.
+ * Interface for determining a device fingerprint for usage within MFA trusted device records.
  *
  * @author Daniel Frett
  * @since 5.3.0
  */
 @FunctionalInterface
-public interface DeviceFingerprintGenerator {
+public interface DeviceFingerprintStrategy {
     /**
-     * Generate a unique browser/device fingerprint for the provided request.
+     * Determine a unique browser/device fingerprint for the provided request.
      *
      * @param principal The principal uid we are generating a fingerprint for.
      * @param context   the request to generate the device fingerprint from.
      * @return The generated fingerprint
      */
-    String generateFingerprint(@Nonnull String principal, @Nonnull RequestContext context);
+    String determineFingerprint(@Nonnull String principal, @Nonnull RequestContext context);
 }
